refactor(notifications-system): Move panel to marketing notifications slot
