+++ js/src/notifications-system/index.js
@@ -1,73 +1,78 @@
 /**
  * External dependencies
  */
-import { createRoot } from '@wordpress/element';
+import { dispatch, resolveSelect, useDispatch } from '@wordpress/data';
+import { createElement } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import NotificationsPanel from './notifications-panel';
-import './index.scss';
-
-const MULTICHANNEL_CLASS = 'woocommerce-marketing-overview-multichannel';
-const BANNER_CLASS = 'woocommerce-marketing-introduction-banner';
-const CONTAINER_CLASS = 'gla-notification-system-container';
-
-let currentRoot = null;
+import { STORE_KEY as GLA_STORE_KEY } from '~/data/constants';
+import { STORE_KEY as MARKETING_NOTIFICATIONS_STORE_KEY } from './woo-marketing-notifications-slot/constants';
+import Notification from './notification';
+import useNotificationsSystemMap from './useNotificationsSystemMap';
 
 /**
- * Mounts the NotificationsPanel into the multichannel marketing overview element.
- * Creates a container and inserts it after the introduction banner, or prepends
- * it if no banner is present. Disconnects the observer once mounted.
+ * Creates a notification component.
  *
- * @param {Element} multichannel The multichannel marketing overview element.
- * @param {MutationObserver} mountObserver The observer to disconnect after mounting.
+ * @param {string} id Notification ID, used to call dismissNotification on dismiss.
+ * @param {number} triggeredAt Unix timestamp (seconds) when the notification was triggered.
+ * @return {Function} A function that returns a React component.
  */
-function mount( multichannel, mountObserver ) {
-	let container = document.querySelector( `.${ CONTAINER_CLASS }` );
+function createNotificationComponent( id, triggeredAt ) {
+	return function NotificationComponent() {
+		const notificationMap = useNotificationsSystemMap();
+		const { dismissNotification } = useDispatch(
+			MARKETING_NOTIFICATIONS_STORE_KEY
+		);
+		const config = notificationMap[ id ];
 
-	if ( container && currentRoot ) {
+		if ( ! config ) {
+			return null;
+		}
+
+		const { title, description, action } = config;
+
+		const handleDismiss = () => {
+			dismissNotification( id );
+		};
+
+		return createElement( Notification, {
+			id,
+			title,
+			description,
+			action,
+			triggeredAt,
+			onDismiss: handleDismiss,
+		} );
+	};
+}
+
+/**
+ * Initializes the notifications system by fetching notifications from the GLA store
+ * and registering them in the marketing notifications store.
+ */
+async function initNotifications() {
+	const glaNotifications =
+		await resolveSelect( GLA_STORE_KEY ).getNotifications();
+
+	if ( ! glaNotifications.length ) {
 		return;
 	}
 
-	if ( ! container ) {
-		container = document.createElement( 'div' );
-		container.className = CONTAINER_CLASS;
+	const { registerNotifications } = dispatch(
+		MARKETING_NOTIFICATIONS_STORE_KEY
+	);
 
-		const banner = multichannel.querySelector( `.${ BANNER_CLASS }` );
+	const notifications = glaNotifications.map( ( { id, triggered_at } ) => {
+		return {
+			id,
+			triggered_at, // The triggered_at timestamp is used for sorting notifications in the marketing notifications store.
+			component: createNotificationComponent( id, triggered_at ),
+		};
+	} );
 
-		if ( banner ) {
-			banner.insertAdjacentElement( 'afterend', container );
-		} else {
-			multichannel.insertBefore( container, multichannel.firstChild );
-		}
-	}
-
-	if ( currentRoot ) {
-		currentRoot.unmount();
-	}
-
-	currentRoot = createRoot( container );
-	currentRoot.render( <NotificationsPanel /> );
-	mountObserver.disconnect();
+	registerNotifications( notifications );
 }
 
-const observer = new MutationObserver( ( _, mountObserver ) => {
-	const multichannel = document.querySelector( `.${ MULTICHANNEL_CLASS }` );
-	if ( multichannel ) {
-		mount( multichannel, mountObserver );
-	}
-} );
-
-observer.observe( document.body, {
-	childList: true,
-	subtree: true,
-} );
-
-const existingMultichannel = document.querySelector(
-	`.${ MULTICHANNEL_CLASS }`
-);
-
-if ( existingMultichannel ) {
-	mount( existingMultichannel, observer );
-}
+initNotifications();
