--- js/src/notifications-system/index.scss
@@ -1,3 +1,0 @@
-.gla-notification-system-container:empty {
-	display: none;
-}
