+++ js/src/notifications-system/woo-marketing-notifications-slot/constants.js
@@ -1,4 +1,4 @@
-export const STORE_NAME = 'wc/marketing-notifications-system';
+export const STORE_KEY = 'wc/marketing-notifications-system';
 export const REGISTER_NOTIFICATIONS = 'REGISTER_NOTIFICATIONS';
 export const DISMISS_NOTIFICATION = 'DISMISS_NOTIFICATION';
 export const MULTICHANNEL_CLASS = 'woocommerce-marketing-overview-multichannel';
