+++ js/src/notifications-system/woo-marketing-notifications-slot/index.scss
@@ -1,0 +1,3 @@
+.woocommerce-marketing-notifications-container:empty {
+	display: none;
+}
