+++ js/src/notifications-system/woo-marketing-notifications-slot/notifications-panel.scss
@@ -1,15 +1,9 @@
-.gla-notifications-panel {
+.gla-woo-marketing-notifications-slot-panel {
 	display: flex;
 	flex-direction: column;
 	gap: $grid-unit-40;
 
-	.gla-notifications-panel__title {
-		align-items: center;
-		display: flex;
-		gap: $grid-unit-20;
-	}
-
-	.gla-notifications-panel__badge {
+	.gla-woo-marketing-notifications-slot-panel__badge {
 		align-items: center;
 		display: flex;
 		font-size: 12px;
@@ -18,4 +12,8 @@
 		min-width: unset;
 		width: 24px;
 	}
+}
+
+.gla-notification-system-container:empty {
+	display: none;
 }
