+++ js/src/notifications-system/woo-marketing-notifications-slot/useNotifications.js
@@ -6,14 +6,14 @@ import { useSelect } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import { STORE_NAME } from './constants';
+import { STORE_KEY } from './constants';
 
 /**
  * @return {{notifications: Array<Object>}} Notifications from the shared marketing notifications store.
  */
 const useNotifications = () => {
 	const notifications = useSelect( ( select ) => {
-		return select( STORE_NAME )?.getNotifications() ?? [];
+		return select( STORE_KEY ).getNotifications();
 	}, [] );
 
 	return { notifications };
