+++ package.json
@@ -176,10 +176,6 @@
 				"path": "./google-listings-and-ads.zip",
 				"maxSize": "8.45 mB",
 				"compression": "none"
-			},
-			{
-				"path": "./js/build/register-notifications.js",
-				"maxSize": "20 kB"
 			}
 		],
 		"ci": {
